+++ askci/apps/main/views/tags.py
@@ -17,8 +17,6 @@ from askci.apps.main.models import Tag
 from askci.apps.main.utils import get_paginated
 from askci.settings import VIEW_RATE_LIMIT as rl_rate, VIEW_RATE_LIMIT_BLOCK as rl_block
 
-import os
-
 
 ## Tag Details